data collector and camera usage updated
The data collector now uses the same processing and cropping settings that I had originally used before going to New Zealand, so let's stick with that since it  requires a lot of effort.

Also great news, I indeed have calibration w.r.t. panda_EE and I think it's a lot more accurate. This is promising ... hopefully in the NEXT commit, we have the ability to convert between pixels and world and vice versa.