Lots of improvements to main.py pipeline
(1) We now save information each trajectory, at each time step, in a more principled way.

(2) We save arguments as well.

(3) We can randomly sample masks for placing now.

(4) Slight tuning of data collection and gripper height thresholding.

(5) We now actually compute the time durations in a more principled way.

We still have to get the z rotations (hurry) but this is a huge administrative coding hurdle.

I also now feel like we could even just use this script for data collection as well... but would be easier to keep that separate.