Tweaks to termination script.
We now add a saving command so that we can save the image data and the metrics data before a CTRL+C and BEFORE we query the GCTN for the action. Think of that as representing the final image and final metric. The reason is that it's hard to get a clean termination condition here and we can just pick ourselves, probably not a big deal. I also made some tweaks to the inspect results script to double check this.